Setup OnceDailySnapshotStrategy
