+++ src/MessageDispatch.Snapshotting.Core.Tests/CheckpointRequestedSnapshotStrategyTests.cs
@@ -17,7 +17,7 @@ public class CheckpointRequestedSnapshotStrategyTests
     [Test]
     public void ShouldSnapshotForEvent_GivenCheckpointRequested_ReturnsTrue()
     {
-        var resolvedEvent = BuildResolvedEvent(EventTypeToSnapshotOn);
+        var resolvedEvent = TestHelpers.BuildResolvedEvent(EventTypeToSnapshotOn);
 
         Assert.That(_strategy.ShouldSnapshotForEvent(resolvedEvent), Is.True);
     }
@@ -26,33 +26,8 @@ public class CheckpointRequestedSnapshotStrategyTests
     [TestCase("CheckpointNotRequested")]
     public void ShouldSnapshotForEvent_GivenCheckpointRequested_ReturnsTrue(string nonSnapshottingEventType)
     {
-        var resolvedEvent = BuildResolvedEvent(nonSnapshottingEventType);
+        var resolvedEvent = TestHelpers.BuildResolvedEvent(nonSnapshottingEventType);
 
         Assert.That(_strategy.ShouldSnapshotForEvent(resolvedEvent), Is.False);
-    }
-
-    private static ResolvedEvent BuildResolvedEvent(string eventType)
-    {
-        var metaData = new Dictionary<string, string>
-        {
-            { "type", eventType },
-            { "created", DateTime.Now.Ticks.ToString() },
-            { "content-type", "application/json" },
-        };
-
-        var serialisedData = JsonSerializer.Serialize(new { Some = "Data" });
-        var serialisedCustomMetaData = JsonSerializer.Serialize(new { Wof = "Tam" });
-
-        var eventRecord = new EventRecord(
-            "event-stream",
-            Uuid.NewUuid(),
-            0,
-            Position.Start,
-            metaData,
-            Encoding.UTF8.GetBytes(serialisedData),
-            Encoding.UTF8.GetBytes(serialisedCustomMetaData));
-
-        var resolvedEvent = new ResolvedEvent(eventRecord, null, null);
-        return resolvedEvent;
     }
 }
